+++ test/landsat_simulation_study/simulation2/simulation_pidx0.8_0.95.R
@@ -68,9 +68,13 @@ for(i in 1:length(pidx0.8_0.95)){
     mat[fidx[j], missing.idx] = NA
 
     #### proposed method
-    res1 <- gapfill_landsat(year, doy, mat, 31, 31,
+    if(file.exists(paste0("./pidx0.8_0.95/res1_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))){
+        res1 <- readRDS(paste0("./pidx0.8_0.95/res1_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+    } else {
+        res1 <- gapfill_landsat(year, doy, mat, 31, 31,
                             use.intermediate.result = FALSE, intermediate.save = FALSE)
-    saveRDS(res1, paste0("./pidx0.8_0.95/res1_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+        saveRDS(res1, paste0("./pidx0.8_0.95/res1_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+    }
     imat = res1$imat[fidx[j],]
     RMSEmat1[j, i] = RMSE(fmat[j, missing.idx], imat[missing.idx])
     NMSEmat1[j, i] = NMSE(fmat[j, missing.idx], imat[missing.idx])
@@ -94,8 +98,13 @@ for(i in 1:length(pidx0.8_0.95)){
     yidxinterval = max(1, yidx - 4):min(16, yidx + 4)
     tmpmat = datarray[,,didxinterval, yidxinterval]
     ## gapfill::Image(tmpmat)
-    res2 = gapfill::Gapfill(tmpmat, clipRange = c(0, 1800), dopar = TRUE)
-    saveRDS(res2, paste0("./pidx0.8_0.95/res2_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+    if(file.exists(paste0("./pidx0.8_0.95/res2_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))){
+        res2 = readRDS(paste0("./pidx0.8_0.95/res2_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+    } else {
+        res2 = gapfill::Gapfill(tmpmat, clipRange = c(0, 1800), dopar = TRUE)
+        saveRDS(res2, paste0("./pidx0.8_0.95/res2_pidx_", pidx0.8_0.95[i], "_fidx_", fidx[j], ".rds"))
+    }
+
     imat = c(res2$fill[,,which(didx == didxinterval), which(yidx == yidxinterval)])
     
     RMSEmat2[j, i] = RMSE(fmat[j, missing.idx], imat[missing.idx])
